requirements.txt
requests
urllib3
pyquery
tldextract
folium
prettytable
ipwhois
dnspython
